Remove documentUnit.ts
RISDEV-8322
