+++ frontend/src/components/DocumentUnitCoreData.vue
@@ -15,7 +15,7 @@ import TitleElement from "@/components/TitleElement.vue"
 import { useValidationStore } from "@/composables/useValidationStore"
 import legalEffectTypes from "@/data/legalEffectTypes.json"
 import { CoreData } from "@/domain/coreData"
-import { Kind } from "@/domain/documentUnit"
+import { Kind } from "@/domain/documentationUnitKind"
 import { pendingProceedingLabels } from "@/domain/pendingProceeding"
 import { SourceValue } from "@/domain/source"
 import ComboboxItemService from "@/services/comboboxItemService"
