+++ frontend/src/components/DocumentUnitSearch.vue
@@ -10,7 +10,7 @@ import Pagination, { Page } from "@/components/Pagination.vue"
 import { useInternalUser } from "@/composables/useInternalUser"
 import { Query } from "@/composables/useQueryFromRoute"
 import { Court } from "@/domain/court"
-import { DocumentUnitSearchParameter } from "@/domain/documentUnit"
+import { DocumentationUnitSearchParameter } from "@/domain/documentationUnitSearchParameter"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import errorMessages from "@/i18n/errors.json"
 import comboboxItemService from "@/services/comboboxItemService"
@@ -27,7 +27,7 @@ const documentUnitListEntries = computed(() => currentPage.value?.content)
 const itemsPerPage = 100
 const searchResponseError = ref()
 const isLoading = ref(false)
-const searchQuery = ref<Query<DocumentUnitSearchParameter>>()
+const searchQuery = ref<Query<DocumentationUnitSearchParameter>>()
 const pageNumber = ref<number>(0)
 const isInternalUser = useInternalUser()
 
@@ -138,9 +138,9 @@ async function updatePage(page: number) {
  * The search form emits an event, when clicking the search button and is
  * triggering the search with the updated query here.
  * It will always reset the pagination to the first page.
- * @param {Query<DocumentUnitSearchParameter>} value - The page to be updated
+ * @param {Query<DocumentationUnitSearchParameter>} value - The page to be updated
  */
-async function updateQuery(value: Query<DocumentUnitSearchParameter>) {
+async function updateQuery(value: Query<DocumentationUnitSearchParameter>) {
   searchQuery.value = value
   pageNumber.value = 0
   await search()
