+++ frontend/src/components/DocumentUnitSearchEntryForm.vue
@@ -9,7 +9,7 @@ import InputField, { LabelPosition } from "@/components/input/InputField.vue"
 import { DropdownItem, ValidationError } from "@/components/input/types"
 import useQuery, { Query } from "@/composables/useQueryFromRoute"
 import { useValidationStore } from "@/composables/useValidationStore"
-import { DocumentUnitSearchParameter } from "@/domain/documentUnit"
+import { DocumentationUnitSearchParameter } from "@/domain/documentationUnitSearchParameter"
 import { PublicationState } from "@/domain/publicationStatus"
 
 defineProps<{
@@ -17,13 +17,13 @@ defineProps<{
 }>()
 
 const emit = defineEmits<{
-  search: [value: Query<DocumentUnitSearchParameter>]
+  search: [value: Query<DocumentationUnitSearchParameter>]
   resetSearchResults: [void]
 }>()
 
-const validationStore = useValidationStore<DocumentUnitSearchParameter>()
+const validationStore = useValidationStore<DocumentationUnitSearchParameter>()
 const { route, getQueryFromRoute, pushQueryToRoute } =
-  useQuery<DocumentUnitSearchParameter>()
+  useQuery<DocumentationUnitSearchParameter>()
 const query = ref(getQueryFromRoute())
 
 const isEmptySearch = computed(() => {
@@ -108,7 +108,7 @@ function resetSearch() {
   emit("resetSearchResults")
 }
 
-function resetErrors(id?: DocumentUnitSearchParameter) {
+function resetErrors(id?: DocumentationUnitSearchParameter) {
   if (id) validationStore.remove(id)
   submitButtonError.value = undefined
 }
@@ -170,9 +170,9 @@ function handleLocalInputError(error: ValidationError | undefined, id: string) {
   if (error) {
     validationStore.add(
       error.message,
-      error.instance as DocumentUnitSearchParameter,
+      error.instance as DocumentationUnitSearchParameter,
     )
-  } else validationStore.remove(id as DocumentUnitSearchParameter)
+  } else validationStore.remove(id as DocumentationUnitSearchParameter)
   validateSearchInput()
 }
 
@@ -252,7 +252,7 @@ watch(
             aria-label="Aktenzeichen Suche"
             fluid
             size="small"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           ></InputText>
         </InputField>
       </div>
@@ -270,7 +270,7 @@ watch(
             fluid
             placeholder="Gerichtstyp"
             size="small"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           ></InputText>
         </InputField>
         <span class="pt-6">-</span>
@@ -287,7 +287,7 @@ watch(
             fluid
             placeholder="Ort"
             size="small"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           ></InputText>
         </InputField>
       </div>
@@ -306,7 +306,7 @@ watch(
             aria-label="Entscheidungsdatum Suche"
             :has-error="hasError"
             @blur="validateSearchInput"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
             @update:validation-error="
               (validationError: ValidationError | undefined) =>
                 handleLocalInputError(validationError, id)
@@ -329,7 +329,7 @@ watch(
             :has-error="hasError"
             placeholder="TT.MM.JJJJ (optional)"
             @blur="validateSearchInput"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
             @update:validation-error="
               (validationError: ValidationError | undefined) =>
                 handleLocalInputError(validationError, id)
@@ -353,7 +353,7 @@ watch(
             :has-error="hasError"
             is-future-date
             @blur="validateSearchInput"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
             @update:validation-error="
               (validationError: ValidationError | undefined) =>
                 handleLocalInputError(validationError, id)
@@ -401,7 +401,7 @@ watch(
             aria-label="Dokumentnummer Suche"
             fluid
             size="small"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           ></InputText>
         </InputField>
       </div>
@@ -421,7 +421,7 @@ watch(
             option-value="value"
             :options="dropdownItems"
             placeholder="Bitte auswählen"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           />
         </InputField>
       </div>
@@ -438,7 +438,7 @@ watch(
             aria-label="Nur meine Dokstelle Filter"
             binary
             :input-id="id"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           />
         </InputField>
         <InputField
@@ -454,7 +454,7 @@ watch(
             aria-label="Nur fehlerhafte Dokumentationseinheiten"
             binary
             :input-id="id"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
           />
         </InputField>
         <InputField
@@ -470,7 +470,7 @@ watch(
             aria-label="Dokumentationseinheiten mit Dublettenverdacht"
             binary
             :input-id="id"
-            @focus="resetErrors(id as DocumentUnitSearchParameter)"
+            @focus="resetErrors(id as DocumentationUnitSearchParameter)"
         /></InputField>
       </div>
       <div class="flex flex-row">
