+++ frontend/src/components/DocumentViewer.vue
@@ -15,7 +15,7 @@ import { useFeatureToggle } from "@/composables/useFeatureToggle"
 import { usePendingProceedingMenuItems } from "@/composables/usePendingProceedingMenuItems"
 import useQuery from "@/composables/useQueryFromRoute"
 import { Decision } from "@/domain/decision"
-import { Kind } from "@/domain/documentUnit"
+import { Kind } from "@/domain/documentationUnitKind"
 import MenuItem from "@/domain/menuItem"
 import PendingProceeding from "@/domain/pendingProceeding"
 import { ResponseError } from "@/services/httpClient"
