+++ frontend/src/components/eurlex/EURLexList.vue
@@ -9,8 +9,7 @@ import DocumentationOfficeSelector from "@/components/DocumentationOfficeSelecto
 import { ComboboxItem } from "@/components/input/types"
 import Pagination, { Page } from "@/components/Pagination.vue"
 import DocumentationOffice from "@/domain/documentationOffice"
-import { EurlexParameters } from "@/domain/documentUnit"
-import EURLexResult from "@/domain/eurlex"
+import EURLexResult, { EurlexParameters } from "@/domain/eurlex"
 import errorMessages from "@/i18n/errors.json"
 import service from "@/services/comboboxItemService"
 import documentationUnitService from "@/services/documentUnitService"
