+++ frontend/src/components/preview/PreviewCoreData.vue
@@ -7,7 +7,8 @@ import PreviewCategory from "@/components/preview/PreviewCategory.vue"
 import PreviewContent from "@/components/preview/PreviewContent.vue"
 import PreviewRow from "@/components/preview/PreviewRow.vue"
 import { CoreData } from "@/domain/coreData"
-import { Kind } from "@/domain/documentUnit"
+
+import { Kind } from "@/domain/documentationUnitKind"
 
 const props = defineProps<{
   coreData: CoreData
