+++ frontend/src/domain/decision.ts
@@ -3,7 +3,7 @@ import ActiveCitation from "@/domain/activeCitation"
 import Attachment from "@/domain/attachment"
 import { ContentRelatedIndexing } from "@/domain/contentRelatedIndexing"
 import { CoreData } from "@/domain/coreData"
-import { Kind } from "@/domain/documentUnit"
+import { Kind } from "@/domain/documentationUnitKind"
 import EnsuingDecision from "@/domain/ensuingDecision"
 import LegalForce from "@/domain/legalForce"
 import { ManagementData } from "@/domain/managementData"
