+++ frontend/src/domain/documentationUnitKind.ts
@@ -1,0 +1,4 @@
+export enum Kind {
+  DOCUMENTION_UNIT = "DOCUMENTION_UNIT",
+  PENDING_PROCEEDING = "PENDING_PROCEEDING",
+}
