+++ frontend/test/components/preview/previewCoreData.spec.ts
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/vue"
 import { previewLayoutInjectionKey } from "@/components/preview/constants"
 import PreviewCoreData from "@/components/preview/PreviewCoreData.vue"
 import { CoreData } from "@/domain/coreData"
-import { Kind } from "@/domain/documentUnit"
+import { Kind } from "@/domain/documentationUnitKind"
 import { SourceValue } from "@/domain/source"
 
 function renderComponent(coreData: CoreData, kind?: Kind) {
